+++ ow_core/Form/Textarea.php
@@ -1,0 +1,86 @@
+<?php
+
+/**
+ * EXHIBIT A. Common Public Attribution License Version 1.0
+ * The contents of this file are subject to the Common Public Attribution License Version 1.0 (the “License”);
+ * you may not use this file except in compliance with the License. You may obtain a copy of the License at
+ * http://www.oxwall.org/license. The License is based on the Mozilla Public License Version 1.1
+ * but Sections 14 and 15 have been added to cover use of software over a computer network and provide for
+ * limited attribution for the Original Developer. In addition, Exhibit A has been modified to be consistent
+ * with Exhibit B. Software distributed under the License is distributed on an “AS IS” basis,
+ * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the specific language
+ * governing rights and limitations under the License. The Original Code is Oxwall software.
+ * The Initial Developer of the Original Code is Oxwall Foundation (http://www.oxwall.org/foundation).
+ * All portions of the code written by Oxwall Foundation are Copyright (c) 2011. All Rights Reserved.
+
+ * EXHIBIT B. Attribution Information
+ * Attribution Copyright Notice: Copyright 2011 Oxwall Foundation. All rights reserved.
+ * Attribution Phrase (not exceeding 10 words): Powered by Oxwall community software
+ * Attribution URL: http://www.oxwall.org/
+ * Graphic Image as provided in the Covered Code.
+ * Display of Attribution Information is required in Larger Works which are defined in the CPAL as a work
+ * which combines Covered Code or portions thereof with code not governed by the terms of the CPAL.
+ */
+
+namespace Oxwall\Core\Form;
+
+/**
+ * Form element: Textarea.
+ *
+ * @author Sardar Madumarov <madumarov@gmail.com>
+ * @since 1.8.3
+ */
+class Textarea extends InvitationFormElement
+{
+
+    /**
+     * Constructor.
+     *
+     * @param string $name
+     */
+    public function __construct( $name )
+    {
+        parent::__construct($name);
+    }
+
+    public function getElementJs()
+    {
+        $jsString = "var formElement = new OwTextArea(" . json_encode($this->getId()) . ", " . json_encode($this->getName()) . ", " . json_encode(( $this->getHasInvitation() ? $this->getInvitation() : false)) . ");";
+
+        /** @var $value Validator  */
+        foreach ( $this->validators as $value )
+        {
+            $jsString .= "formElement.addValidator(" . $value->getJsValidator() . ");";
+        }
+
+        return $jsString;
+    }
+
+    /**
+     * @see FormElement::renderInput()
+     *
+     * @param array $params
+     * @return string
+     */
+    public function renderInput( $params = null )
+    {
+        parent::renderInput($params);
+
+        if ( $this->getValue() === null && $this->getHasInvitation() )
+        {
+            $this->addAttribute("value", $this->getInvitation());
+            $this->addAttribute("class", "invitation");
+        }
+        else if ( $this->getValue() !== null )
+        {
+            $this->addAttribute("value", $this->getValue());
+        }
+
+        $content = $this->getAttribute("value");
+        $this->removeAttribute("value");
+
+        $markup = \Oxwall\Utilities\HtmlTag::generateTag("textarea", $this->attributes, true, $content);
+
+        return $markup;
+    }
+}
