+++ ow_core/Form/UrlValidator.php
@@ -22,14 +22,33 @@
  * which combines Covered Code or portions thereof with code not governed by the terms of the CPAL.
  */
 
+namespace Oxwall\Core\Form;
+
+use Oxwall\Core\OW;
+
 /**
- * Base form class.
- * 
- * @author Sardar Madumarov <madumarov@gmail.com>
- * @package ow_core
- * @since 1.0
+ * UrlValidator validates Url.
+ *
+ * @author Podyachev Evgeny <joker.OW2@gmail.com>
+ * @since 1.8.3
  */
-class Form extends Oxwall\Core\Form\Form
+class UrlValidator extends RegExpValidator
 {
-    
+
+    /**
+     * Class constructor
+     */
+    public function __construct()
+    {
+        parent::__construct(\Oxwall\Utilities\Validator::URL_PATTERN);
+
+        $errorMessage = OW::getLanguage()->text('base', 'form_validator_url_error_message');
+
+        if ( empty($errorMessage) )
+        {
+            $errorMessage = 'Url Validator Error!';
+        }
+
+        $this->setErrorMessage($errorMessage);
+    }
 }
